Transformation/function_bus_tranformation_requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=2.0.0
google-cloud-storage
pandas
pyarrow